AST: promote AList to module; Haddock TOC/ordering
AList declares a fromList function, which preferably doesn't go into the
global namespace. Upgrading it to a module gives us the flexibility to
change this easily if it becomes a user issue.

SpannedPair and generic Spanned, SpannedPair instances have been moved
from AST to Util.Position . (Without this, the AList module has a cyclic
dependency on AST due to being re-exported.)

The AST module is now sectioned in the Haddock. Easier to parse.
