+++ fortran-src.cabal
@@ -4,7 +4,7 @@ cabal-version: 1.12
 --
 -- see: https://github.com/sol/hpack
 --
--- hash: 50aa4fb72aab53b1e5d449fa390ff3290fb26bd9bf9297a7cac174a79ef92b3f
+-- hash: 859eadbf578a446779b4602ab015d0a4d996ec983f0265dcf81b31aafd01c137
 
 name:           fortran-src
 version:        0.4.2
@@ -35,6 +35,7 @@ library
       Language.Fortran.Analysis.BBlocks
       Language.Fortran.Analysis.DataFlow
       Language.Fortran.AST
+      Language.Fortran.AST.AList
       Language.Fortran.Version
       Language.Fortran.LValue
       Language.Fortran.Intrinsics
