+++ package.yaml
@@ -52,6 +52,7 @@ library:
   - Language.Fortran.Analysis.BBlocks
   - Language.Fortran.Analysis.DataFlow
   - Language.Fortran.AST
+  - Language.Fortran.AST.AList
   - Language.Fortran.Version
   - Language.Fortran.LValue
   - Language.Fortran.Intrinsics
